paquo.objects.classes: factor out QuPathPathClass
